CrossPlatform: Fix boost implementation of Wait
When executing Wait, without having executed Reset,
the implementation used to erroneously return false,
even though the object was already signaled.

This lead to the AbkClient badly assuming that Event Data has not been read, yet.

Furthermore, the object is created in manual-reset mode under Win32,
the auto-reset of m_eventCond does not reflect this. It has thus been removed.